+++ misc/pstree.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+# echo $BASHPID > /tmp/pid
+# watch -n1 pstree.sh
+pstree -U --hide-threads -c -a `cat /tmp/pid`
+echo
+find /tmp -maxdepth 1 -name '*.vilarktmp'