+++ src/controllers/Main.cs
@@ -9,9 +9,10 @@ class Controller
 {
     // Models / Data
     private Console console;
-    private EventQueue<KeyPress> keyboardEvent;
-    private EventQueue<SignalPayload> signalEvent;
-    private EventQueue<LoadProgressInfo> loadingEvent;
+    private EventQueue<KeypressPayload> m_keyboard_events;
+    private EventQueue<SignalPayload> m_signal_events;
+    public EventQueue<Notification> m_notifications;
+    public EventQueue<string> m_web_replies;
     private Config m_config;
     private InputModel m_input_model;
     private OutputModel m_output_model;
@@ -25,24 +26,23 @@ class Controller
     private BottomBar m_bottombar;
 
     // User / UX State
-    private IScrollItem? m_chosen_item = null;
-    private bool m_quit_signaled = false;
+    private EventWaitHandle? m_keyboard_paused = null;
+    private bool        m_child_process_running = false;
+    private bool        m_web_request_running = false;
     private System.Timers.Timer? m_redraw_timer = null;  // For the loading spinner only
 
-    public Controller(Console console,
-                            EventQueue<KeyPress> keyboardEvent,
-                            EventQueue<SignalPayload> signalEvent,
-                            OptionsModel optionsModel)
+    public Controller(Console console, OptionsModel optionsModel)
     {
         this.console = console;
-        this.keyboardEvent = keyboardEvent;
-        this.signalEvent = signalEvent;
-        this.loadingEvent = new();
+        m_keyboard_events = new();
+        m_signal_events = new();
+        m_notifications = new();
+        m_web_replies = new();
         m_options_model = optionsModel;
 
         m_config = new();
-        m_input_model = new(optionsModel, loadingEvent);
-        m_output_model = new();
+        m_input_model = new(optionsModel, m_notifications);
+        m_output_model = new(m_config);
 
         // Views
         m_titlebar = new();
@@ -63,24 +63,11 @@ class Controller
     }
 
     public void Run(string[] args) {
-        console.SetAlternateScreen(true);
-        if (Environment.GetEnvironmentVariable("VILARK_NOCLEAR") == null) {
-            console.ClearScreen();
-        }
-        console.Flush();
-
         try {
             PrepareViews();
             Redraw();
+            m_input_model.StartLoadingAsync();
             RunEventLoop();
-            Log.Info("Clean exit");
-
-            CleanupTerminal();
-            m_config.SaveSettings();
-
-            // This writes a tmp file, or execs a process
-            m_output_model.WriteOutput(m_chosen_item);
-            Environment.Exit(0);
         } catch (Exception e) {
             Log.Info("Unclean exit");
             Log.Exception(e);
@@ -92,6 +79,7 @@ class Controller
     }
 
     private void CleanupTerminal() {
+        Log.Info($"CleanupTerminal()");
         console.SetCursorVisible(true);
         console.SetUnderline(false);
         console.ResetTextAttrs();
@@ -107,35 +95,34 @@ class Controller
     }
 
     private void RunEventLoop() {
-
-        m_input_model.StartLoadingAsync();
-
         var waitHandles = new WaitHandle[] {
-            keyboardEvent.ConsumerWaitHandle,
-            signalEvent.ConsumerWaitHandle,
-            loadingEvent.ConsumerWaitHandle
+            m_keyboard_events.ConsumerWaitHandle,
+            m_signal_events.ConsumerWaitHandle,
+            m_notifications.ConsumerWaitHandle
         };
         while (true) {
             int whichReady = WaitHandle.WaitAny(waitHandles);
             if (whichReady == 0) {
                 // Keyboard
-                KeyPress kp = keyboardEvent.TakeEvent();
-                onKeyPress(kp);
-                if (m_quit_signaled) {
-                    return;
+                KeypressPayload payload = m_keyboard_events.TakeEvent();
+                onKeyPress(payload.keypress);
+                if (!m_child_process_running || m_web_request_running) {
+                    // Allow keyboard thread to read again from tty
+                    payload.doneProcessing.Set();
+                } else {
+                    // Don't reenable keyboard yet.  We still need to respond
+                    // to signals like ctrl-z however.
+                    m_keyboard_paused = payload.doneProcessing;
                 }
             } else if (whichReady == 1) {
                 // Posix Signal
-                SignalPayload payload = signalEvent.TakeEvent();
+                SignalPayload payload = m_signal_events.TakeEvent();
                 onPosixSignal(payload.signal);
-                if (m_quit_signaled) {
-                    return;
-                }
                 // Allow the waiting signal handler thread to return
                 payload.doneProcessing.Set();
             } else if (whichReady == 2) {
-                LoadProgressInfo progress = loadingEvent.TakeEvent();
-                onProgressEvent(progress);
+                Notification notification = m_notifications.TakeEvent();
+                onNotification(notification);
             } else {
                 throw new Exception($"invalid whichReady {whichReady}");
             }
@@ -162,11 +149,14 @@ class Controller
     }
 
     private void Redraw() {
+        Log.Info("Redraw()");
         UpdateViewDimensions();
+
+        // Ensure we always draw on alternate screen.
+        console.SetAlternateScreen(true);
 
         // Hide cursor while redrawing
         console.SetCursorVisible(false);
-        console.Flush();
 
         m_titlebar.DrawIfVisible(console);
         m_main_tab.DrawIfVisible(console);
@@ -184,9 +174,8 @@ class Controller
 
     private void OnRedrawTimer(object? sender, ElapsedEventArgs args) {
         // NB: This is called on a threadpool thread.
-        // Fake a SIGWINCH signal just to force a redraw on the main thread
         Log.Info("OnRedrawTimer Start");
-        ExternalSignalInput(PosixSignal.SIGWINCH);
+        m_notifications.AddEvent(new Notification(ForceRedraw: true));
         Log.Info("OnRedrawTimer End");
     }
 
@@ -195,8 +184,7 @@ class Controller
     }
 
     private void OnItemChosen(object? sender, IScrollItem item) {
-        m_chosen_item = item;
-        m_quit_signaled = true;
+        DoChosenItem(item);
     }
 
     private void UpdateSearchModel() {
@@ -224,7 +212,12 @@ class Controller
         Log.Info($"got keypress {kp}");
 
         if (kp.keyCode == KeyCode.ESCAPE) {
-            m_quit_signaled = true; // Exit
+            if (m_web_request_running) {
+                m_web_replies.AddEvent("");
+                m_web_request_running = false;
+            } else {
+                DoExit();
+            }
         } else if (kp.keyCode == KeyCode.RIGHT_ARROW) {
             m_titlebar.CycleTabs(1);
             ActiveTabChanged();
@@ -237,41 +230,48 @@ class Controller
             m_options_tab.OnKeyPress(kp);
         }
 
-        // We redraw after every keypress, no need to handle Ctrl-l specifically
-        Redraw();
+        // We redraw after every keypress, no need to handle Ctrl-L specifically
+        if (!m_child_process_running || m_web_request_running) {
+            Redraw();
+        }
     }
 
     private void onPosixSignal(PosixSignal sig) {
-        Log.Info($"onPosixSignal {sig}");
+        Log.Info($"onPosixSignal {sig}, child_running={m_child_process_running}, web_running={m_web_request_running}");
 
         if (sig == PosixSignal.SIGWINCH || sig == PosixSignal.SIGCONT) {
-            console.SetAlternateScreen(true);
-            Redraw();
+            if (!m_child_process_running || m_web_request_running) {
+                Redraw();
+            }
         }
+
         if (sig == PosixSignal.SIGINT) {
-            Log.Info("moving back to main terminal screen (SIGINT)");
+            // The default handler will terminate us when this func returns
             CleanupTerminal();
         }
+
         if (sig == PosixSignal.SIGTSTP) {
-            Log.Info("moving back to main terminal screen (SIGTSTP)");
             CleanupTerminal();
             // The default action of SIGTSTP doesn't send a STOP when we capture it.
             // So send it ourself.
             UnixProcess.SelfSigStop();
         }
+
         if (sig == PosixSignal.SIGTERM) {
-            m_quit_signaled = true;
+            DoExit();
         }
     }
 
-    private void onProgressEvent(LoadProgressInfo progress) {
-        Log.Info($"Got progress {progress}");
-        m_main_tab.m_loading_view.CurrentData = progress;
-        if (progress.CompletedData != null) {
+    private void onNotification(Notification notification) {
+        Log.Info($"Got notification {notification}");
+        if (notification.Processed != null) {
+            m_main_tab.m_loading_view.CurrentData = notification;
+        }
+        if (notification.CompletedData != null) {
             // Files all done loading.  Turn off the progress bar, and move to main UX
             m_main_tab.m_loading_view.SetVisible(false);
             m_main_tab.m_scrollview.SetVisible(true);
-            m_input_model.SetCompletedData(progress.CompletedData);
+            m_input_model.SetCompletedData(notification.CompletedData);
             // Update in case the user already typed something
             UpdateSearchModel();
 
@@ -280,11 +280,61 @@ class Controller
 
             Redraw();
         }
-        if (progress.ErrorMessage != null) {
+        if (notification.ErrorMessage != null) {
             CleanupTerminal();
             System.Console.WriteLine("Error when scanning files");
-            System.Console.WriteLine(progress.ErrorMessage);
+            System.Console.WriteLine(notification.ErrorMessage);
             Environment.Exit(1);
+        }
+        if (notification.ChildExited == true) {
+            Log.Info("Child process exited, re-enabling keyboard / tty reads.");
+            m_child_process_running = false;
+            EnableKeyboard();  // No reason to be suspended
+            Redraw();
+        }
+        if (notification.ForceRedraw) {
+            // Sent by the loading progress timer
+            Redraw();
+        }
+        if (notification.WebRequest != null) {
+            Log.Info("Fast switching to our UX, for IPC request");
+            m_web_request_running = true;
+            if (m_config.FastSwitchSearch == FastSwitchSearch.FAST_CLEAR_SEARCH) {
+                m_main_tab.m_searchbar.ClearSearch();
+            }
+            EnableKeyboard();
+            Redraw();
+        }
+    }
+
+    private void DoExit() {
+        Log.Info($"Exiting due to escape or SIGTERM. Nothing chosen. child={m_child_process_running}");
+        CleanupTerminal();
+        m_config.SaveSettings();
+        m_output_model.WriteOutput(null);
+        Environment.Exit(0);
+    }
+
+    private void DoChosenItem(IScrollItem item) {
+        Log.Info("An item was chosen.");
+
+        // This will either:
+        // a. execve() and not return (it could throw an exception on failure)
+        // b. Process.Start() and wait for child process to exit
+        if (m_output_model.GetEditorCommand() != null) {
+            if (m_web_request_running) {
+                var fullPath = Path.GetFullPath(item.GetSelectionString());
+                m_web_replies.AddEvent(fullPath);
+                m_web_request_running = false;
+            } else {
+                m_output_model.LaunchEditor(item, m_notifications);
+                m_child_process_running = true;
+            }
+        } else {
+            CleanupTerminal();
+            m_config.SaveSettings();
+            m_output_model.WriteOutput(item);
+            Environment.Exit(0);
         }
     }
 
@@ -294,8 +344,24 @@ class Controller
     public void ExternalSignalInput(PosixSignal signal) {
         using (var doneProcessing = new EventWaitHandle(false, EventResetMode.AutoReset)) {
             var payload = new SignalPayload(signal, doneProcessing);
-            signalEvent.AddEvent(payload);
+            m_signal_events.AddEvent(payload);
             payload.doneProcessing.WaitOne();
+        }
+    }
+
+    // Called by a separate console read thread.
+    // This does not return until our main thread ack's it.
+    // This is so we can choose to launch a child process like vim and wait for it,
+    // and not read from the tty while it is running.
+    public void ExternalKeyboardInput(KeypressPayload payload) {
+        m_keyboard_events.AddEvent(payload);
+    }
+
+    private void EnableKeyboard() {
+        if (m_keyboard_paused != null) {
+            Log.Info("Enabling keyboard");
+            m_keyboard_paused.Set();
+            m_keyboard_paused = null;
         }
     }