+++ src/main/Program.cs
@@ -1,13 +1,14 @@
 // Copyright (C) 2023 Karl Pickett / ViLark Project
-using System.Collections;
 using System.Runtime.InteropServices;
+using System.Collections;
 using vilark;
 
+record struct KeypressPayload(KeyPress keypress, EventWaitHandle doneProcessing);
 record struct SignalPayload(PosixSignal signal, EventWaitHandle doneProcessing);
 
 class ViLarkMain
 {
-    public const string VERSION = "1.3";
+    public const string VERSION = "1.5";
 
     static void Main(string[] args)
     {
@@ -32,12 +33,10 @@ class ViLarkMain
         // Parse CLI options.  This will exit if anything is invalid.
         var optionsModel = new OptionsModel();
 
-        // Options look ok, start up the UX
+        // Options look ok, continue startup.
+        // Set up the UX and signal handlers.
         var console = new vilark.Console();
-        var keyboard = new Keyboard();
-        var keyboardEvent = new EventQueue<KeyPress>();
-        var signalEvent = new EventQueue<SignalPayload>();
-        var controller = new Controller(console, keyboardEvent, signalEvent, optionsModel);
+        var controller = new Controller(console, optionsModel);
 
         // Synchronization is important here.. one signal at a time is processed by main thread,
         // and we wait until it is done before returning to the runtime.
@@ -76,21 +75,36 @@ class ViLarkMain
         // This is terrible if left at the default, it kills us and leaves us on the app screen.
         // So we catch it, do basic terminal cleanup, and exit.
         //
-        // The default action seems to race with our handler (50/50), so we cancel it.
-        //
         // I don't know of many full screen apps that do exit on ctrl-c (tig does), but
         // we're just a simple dialog so it's probably fine.
         UnixProcess.RegisterSignalHandler(PosixSignal.SIGINT, (context) => captureSignal(context, false));
 
         // Read from keyboard/tty in a separate thread.
+        var keyboard = new Keyboard();
         var consoleReadThread = new Thread(() => {
+                var doneProcessing = new EventWaitHandle(false, EventResetMode.AutoReset);
                 foreach (KeyPress kp in keyboard.GetKeyPress()) {
-                    keyboardEvent.AddEvent(kp);
+                    controller.ExternalKeyboardInput(new KeypressPayload(kp, doneProcessing));
+                    doneProcessing.WaitOne();
                 }
             });
         consoleReadThread.Name = "consoleReadThread";
         consoleReadThread.Start();
 
+        // For super fast IPC, without launching another process :-)
+        if (Environment.GetEnvironmentVariable("VILARK_IPC_URL") == null) {
+            var web = new WebListener(controller);
+            Environment.SetEnvironmentVariable("VILARK_IPC_URL", web.GetUrl());
+            var webThread = new Thread(() => {
+                    web.Run();
+                });
+            webThread.Name = "webThread";
+            webThread.Start();
+        } else {
+            Log.Info("VILARK_IPC_URL already set, not starting another web listener");
+        }
+
+        // Run eventloop
         controller.Run(args);
     }